% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rStagger.R
\name{callStagger.parallel}
\alias{callStagger.parallel}
\title{callStagger.parallel}
\usage{
callStagger.parallel(files, file.dir, out.dir, jarFile = NULL,
  modelFile = NULL)
}
\arguments{
\item{files}{A vector of textfiles (file names)}

\item{file.dir}{Path to directory of textfiles}

\item{out.dir}{Path to directory for output files}

\item{jarFile}{Path to .jar (Stagger) file. Default NULL (uses included .jar file)}

\item{modelFile}{Path to .bin modelfile (i.e. the Swedish file). Default NULL (uses included modelfile)}
}
\value{
Returns a .txt file containing the parsed text
}
\description{
A function to call Stagger (NLP) from R in parallel.  Works currently only on a UNIX system with Java installed.
}
\details{
Uses GNU Parallel. See more at \href{https://www.gnu.org/software/parallel/}{https://www.gnu.org/software/parallel/}
}
\examples{
\donttest{callStagger.parallel(files = c("test1.txt", "test2.txt"), file.dir = "~/Desktop/", "~/Desktop/")}

}
