% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rStagger.R
\name{callStagger}
\alias{callStagger}
\title{callStagger}
\usage{
callStagger(textFile, jarFile = NULL, modelFile = NULL)
}
\arguments{
\item{textFile}{String or path to .txt file. If string, the string will be written to a temporary file that is deleted once the algorithm finishes}

\item{jarFile}{Path to .jar (Stagger) file. Default NULL (uses included .jar file)}

\item{modelFile}{Path to .bin modelfile (i.e. the Swedish file). Default NULL (uses included modelfile)}
}
\value{
Returns a dataframe with class c("data.frame", "rStagger")
}
\description{
A function to call Stagger (NLP) from R
}
