% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rStagger.R
\name{countPosTags}
\alias{countPosTags}
\title{countPosTags}
\usage{
countPosTags(object)
}
\arguments{
\item{object}{A rStagger object}
}
\value{
Returns a dataframe with POS frequencies c("data.frame", "rStagger", "rStaggerCount")
}
\description{
A function to count the frequencies of POS tags
}
