% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rStagger.R
\name{countTerm}
\alias{countTerm}
\title{countTerm}
\usage{
countTerm(object, type = c("Token", "Lemma"))
}
\arguments{
\item{object}{A rStagger object}

\item{type}{Term type, Tokens or Lemmas}
}
\value{
Returns a dataframe with Token or Lemma frequencies c("data.frame", "rStagger", "rStaggerCount")
}
\description{
A function to count the frequencies of Tokens or Lemmas
}
