% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{dictonary}
\alias{dictonary}
\title{dictonary}
\usage{
dictonary(type = c("POS"))
}
\arguments{
\item{type}{Type of dictionary. Currently only Part-of-speech dictionary}
}
\value{
A lookup vector
}
\description{
Get dictionary of various tags
}
