% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rStagger.R
\name{extract_entity}
\alias{extract_entity}
\title{extract_entity}
\usage{
extract_entity(object, type = c("all", "person", "place", "inst", "product",
  "myth", "work", "animal", "other"))
}
\arguments{
\item{object}{An object of class rStagger}

\item{type}{Entity type to extract: c("all","person","place","inst","product","myth","work","animal","other")}
}
\value{
Returns a string of entities (the chosen entity type)
}
\description{
A function to extract entities from a rStagger object
}
