% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{extract_features}
\alias{extract_features}
\title{extract_features}
\usage{
extract_features(object, type = c("Readability", "Stylometry"))
}
\arguments{
\item{object}{A rStagger object}

\item{type}{Type of text feature to extract. Current choices c("Readability", "Stylometry")}
}
\value{
A vector of features
}
\description{
A function to extract various measures from an rStagger object from \code{\link{callStagger}} or \code{\link{readParsed}}
}
\details{
The following features are implemented
 \itemize{
  \item Stylometry
  \itemize{
  \item{TTR: Type-Token Ratio}
  \item{DL: Document length as in number of words}
  \item{HAPAX: Hapax Legomenon}
}
  \item Readability
  \itemize{
  \item{LIX: Readability index (Swedish: Lasbarhetsindex)}
  \item{MSL: Mean sentence length calculated as the average number of words per sentence}
  \item{MWLC: Mean word length calculated as the average number of characters per word}
}
}
}
