% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rStagger.R
\name{readParsed}
\alias{readParsed}
\title{readParsed}
\usage{
readParsed(file, ...)
}
\arguments{
\item{file}{Path to file}

\item{...}{Other argument to pass to internal calls (read.delim())}
}
\value{
Returns a dataframe with class c("data.frame", "rStagger")
}
\description{
A function to read Stagger output into data.frame
}
