\name{rStagger-package}
\alias{rStagger-package}
\alias{rStagger}
\docType{package}
\title{
\packageTitle{rStagger}
}
\description{
\packageDescription{rStagger}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rStagger}
\packageIndices{rStagger}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{rStagger}

Maintainer: \packageMaintainer{rStagger}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
