% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rStagger.R
\name{stringToDF}
\alias{stringToDF}
\title{stringToDF}
\usage{
stringToDF(string, ...)
}
\arguments{
\item{string}{A string file}

\item{...}{Other argument to pass to internal calls (read.delim())}
}
\description{
Utility function
}
