% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{readParsed}
\alias{readParsed}
\title{readParsed}
\usage{
readParsed(file, ...)
}
\arguments{
\item{file}{Path to file}

\item{...}{Other argument to pass to internal calls (read.delim())}
}
\value{
Returns a dataframe with class c("data.frame", "rStagger")
}
\description{
A function to read Stagger output into data.frame
}
\details{
Based the CoNLL-X annotated data format. 
See more at \href{https://ilk.uvt.nl/~emarsi/download/pubs/14964.pdf}{https://ilk.uvt.nl/~emarsi/download/pubs/14964.pdf}
}
