\name{rStagger-package}
\alias{rStagger-package}
\alias{rStagger}
\docType{package}
\title{
\packageTitle{rStagger}
}
\description{
\packageDescription{rStagger}
}
\author{
\packageAuthor{rStagger}

Maintainer: \packageMaintainer{rStagger}
}
\references{
Stagger: \href{https://www.ling.su.se/english/nlp/tools/stagger}{https://www.ling.su.se/english/nlp/tools/stagger}  
  

GNU Parallel: \href{https://www.gnu.org/software/parallel/}{https://www.gnu.org/software/parallel/}
}
\keyword{NLP, Stagger, }
